+++ src/correlator_node_data_reader_tasklet.cc
@@ -26,7 +26,7 @@ get_output_buffer() {
 
 void Correlator_node_data_reader_tasklet::do_task() {
   SFXC_ASSERT(has_work());
-  size_t n_bytes; 
+  int32_t n_bytes; 
   int n_invalid_blocks,n_delays;
 
   Output_memory_pool_element output_elem = output_memory_pool.allocate();
